+++ src/ng-select/ng-select.component.ts
@@ -31,7 +31,8 @@ import {
     NgOptionTemplateDirective,
     NgLabelTemplateDirective,
     NgHeaderTemplateDirective,
-    NgFooterTemplateDirective
+    NgFooterTemplateDirective,
+    NgOptgroupTemplateDirective
 } from './ng-templates.directive';
 
 import { NgOption, KeyCode, NgSelectConfig } from './ng-select.types';
@@ -59,7 +60,6 @@ export type DropdownPosition = 'bottom' | 'top' | 'auto';
         '[class.top]': 'currentDropdownPosition === "top"',
         '[class.bottom]': 'currentDropdownPosition === "bottom"',
         '[class.ng-single]': '!multiple',
-        '[class.ng-selected]': 'hasValue'
     }
 })
 export class NgSelectComponent implements OnDestroy, OnChanges, AfterViewInit, ControlValueAccessor {
@@ -81,6 +81,7 @@ export class NgSelectComponent implements OnDestroy, OnChanges, AfterViewInit, C
     @Input() loading = false;
     @Input() closeOnSelect = true;
     @Input() maxSelectedItems: number;
+    @Input() groupBy: string;
     @Input() bufferAmount = 4;
     @Input() virtualScroll = false;
     @Input() @HostBinding('class.typeahead') typeahead: Subject<string>;
@@ -102,6 +103,7 @@ export class NgSelectComponent implements OnDestroy, OnChanges, AfterViewInit, C
 
     // custom templates
     @ContentChild(NgOptionTemplateDirective, { read: TemplateRef }) optionTemplate: TemplateRef<any>;
+    @ContentChild(NgOptgroupTemplateDirective, { read: TemplateRef }) optgroupTemplate: TemplateRef<any>;
     @ContentChild(NgLabelTemplateDirective, { read: TemplateRef }) labelTemplate: TemplateRef<any>;
     @ContentChild(NgHeaderTemplateDirective, { read: TemplateRef }) headerTemplate: TemplateRef<any>;
     @ContentChild(NgFooterTemplateDirective, { read: TemplateRef }) footerTemplate: TemplateRef<any>;
@@ -134,8 +136,8 @@ export class NgSelectComponent implements OnDestroy, OnChanges, AfterViewInit, C
         this.unselect(option);
     };
 
-    constructor(@Inject(NG_SELECT_DEFAULT_CONFIG) config: NgSelectConfig,
-        private changeDetectorRef: ChangeDetectorRef,
+    constructor( @Inject(NG_SELECT_DEFAULT_CONFIG) config: NgSelectConfig,
+        private _cd: ChangeDetectorRef,
         public elementRef: ElementRef
     ) {
         this._mergeGlobalConfig(config);
@@ -362,7 +364,7 @@ export class NgSelectComponent implements OnDestroy, OnChanges, AfterViewInit, C
         return empty && this._isTypeahead && !this.filterValue && !this.isLoading;
     }
 
-    onFilter(term: string) {
+    filter(term: string) {
         if (!this.searchable) {
             return;
         }
@@ -399,8 +401,8 @@ export class NgSelectComponent implements OnDestroy, OnChanges, AfterViewInit, C
     }
 
     detectChanges() {
-        if (!(<any>this.changeDetectorRef).destroyed) {
-            this.changeDetectorRef.detectChanges();
+        if (!(<any>this._cd).destroyed) {
+            this._cd.detectChanges();
         }
     }
 
@@ -450,7 +452,7 @@ export class NgSelectComponent implements OnDestroy, OnChanges, AfterViewInit, C
                 .subscribe(option => {
                     const item = this.itemsList.findItem(option.value);
                     item.disabled = option.disabled;
-                    this.changeDetectorRef.markForCheck();
+                    this._cd.markForCheck();
                 });
         }
 
@@ -531,7 +533,7 @@ export class NgSelectComponent implements OnDestroy, OnChanges, AfterViewInit, C
         }
         this._ngModel = ngModel;
         this.changeEvent.emit(this._value);
-        this.changeDetectorRef.markForCheck();
+        this._cd.markForCheck();
     }
 
     private _clearSearch() {